WIP CreditSystem integration
CreditSystem shared pointer is created on the Unreal side and propogated as desired to the Cesium Native credit aggregation code. TODO: Move the resulting list of credits from ViewUpdateResult to the tileset's CreditSystem. TODO: Move the creation of the CreditSystem to the point of construction of ACesiumTileset objects so the same CreditSystem pointer can be passed to multiple tilesets.